Move contexts to context sub-folder
